requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
pyaml==17.12.1
PyYAML==3.12
requests==2.18.4
urllib3==1.22
wheel==0.24.0